# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pulp API",
    "description": "Pulp social-listening HTTP API. This spec is generated from the Rust backend and is the authoritative API contract.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/logs/{service}": {
      "get": {
        "tags": [
          "logs"
        ],
        "summary": "`GET /api/logs/{service}` — last N log lines for `service`.",
        "description": "`service` accepts any channel name (`reddit`, `hackernews`, `github`) and\nfuture services (`ai_filter`/`llm`). Unknown services → 404. The number of\nlines is capped (default 200, max 2000).",
        "operationId": "getServiceLogs",
        "parameters": [
          {
            "name": "service",
            "in": "path",
            "description": "Service id: a channel name or ai_filter/llm",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Max lines to return (default 200, max 2000)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown service"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LogResponse": {
        "type": "object",
        "description": "Response body: the resolved service, the lines (oldest → newest), and\nwhether the underlying log file existed.",
        "required": [
          "service",
          "lines",
          "exists"
        ],
        "properties": {
          "exists": {
            "type": "boolean"
          },
          "lines": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}